third_party/blink/web_tests/fast/frames/iframe-no-src-set-location-expected.txt

This tests that an iframe with no src attribute uses the parent document's URL as a base when resolving non-fully-qualified URLs set by JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('theFrame').contentDocument.baseURI is document.baseURI
PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById('theFrame').contentDocument.body.innerHTML is 'PASS\n'
PASS successfullyParsed is true

TEST COMPLETE
